Add files via upload
This is an Excel macro created using the VBA language to automatic an manual process of copying the average daily traffic counts from the raw Excel spreadsheet to a master list (Excel spreadsheet). This macro loops through many raw Excel files in a folder, can detect where of two possible cell locations the traffic count number resides in the raw Excel spreadsheet, uses the raw file name (with the help of a string function to strip out the file extension) as the count location id and finds which row in the master list this station id is located and places the count number in the last column of its corresponding row location, and then formats the number in using a thousands separator and no decimal places.